requirements.txt
# ── Core Framework ─────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
pydantic>=2.0

# ── Authentication ─────────────────────────────────
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# ── LangChain + RAG ───────────────────────────────
langchain>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0
langchain-text-splitters>=0.3.0
langchain-core>=0.3.0

# ── Document Loaders ──────────────────────────────
pdfplumber>=0.11.0
pypdf>=5.0.0

# ── Embeddings ────────────────────────────────────
sentence-transformers>=3.0.0

# ── Vector Store ──────────────────────────────────
chromadb>=0.6.0

# ── LLM Backend ──────────────────────────────────
ollama>=0.4.0

# ── Environment ──────────────────────────────────
python-dotenv>=1.0.0

# ── Testing ──────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
